refactor: 해시태그 값을 입력해서 데이터를 전송할 수 있도록 변경
